Add files via upload
Added more documentation. Rings and trinkets will now un-equip how they should. Before they only unequipped the array[10] or array[11] now they will look at both slots for rings which is array[10] and array[11] and for trinkets array[12] and array[13]. Also fixed an error in the equipgear method that was searching for item objects on empty gear slots. It no longer does that.